Move Setting, RolloutRule, SettingType, PercentageRule to public scope
