+++ src/main/java/com/configcat/ConfigModels.java
@@ -6,13 +6,6 @@ import com.google.gson.annotations.SerializedName;
 import java.util.HashMap;
 import java.util.Map;
 
-enum SettingType {
-    BOOLEAN,
-    STRING,
-    INT,
-    DOUBLE,
-}
-
 class Entry {
     @SerializedName(value = "c")
     private Config config;
@@ -87,93 +80,3 @@ class Preferences {
     }
 }
 
-class Setting {
-    @SerializedName(value = "v")
-    private JsonElement value;
-    @SerializedName(value = "t")
-    private int type;
-    @SerializedName(value = "p")
-    private PercentageRule[] percentageItems;
-    @SerializedName(value = "r")
-    private RolloutRule[] rolloutRules;
-    @SerializedName(value = "i")
-    private String variationId = "";
-
-    public void setValue(JsonElement value) {
-        this.value = value;
-    }
-
-    public JsonElement getValue() {
-        return value;
-    }
-
-    public int getType() {
-        return type;
-    }
-
-    public PercentageRule[] getPercentageItems() {
-        return percentageItems;
-    }
-
-    public RolloutRule[] getRolloutRules() {
-        return rolloutRules;
-    }
-
-    public String getVariationId() {
-        return variationId;
-    }
-}
-
-class RolloutRule {
-    @SerializedName(value = "v")
-    private JsonElement value;
-    @SerializedName(value = "a")
-    private String comparisonAttribute;
-    @SerializedName(value = "t")
-    private int comparator;
-    @SerializedName(value = "c")
-    private String comparisonValue;
-    @SerializedName(value = "i")
-    private String variationId;
-
-    public JsonElement getValue() {
-        return value;
-    }
-
-    public String getComparisonAttribute() {
-        return comparisonAttribute;
-    }
-
-    public int getComparator() {
-        return comparator;
-    }
-
-    public String getComparisonValue() {
-        return comparisonValue;
-    }
-
-    public String getVariationId() {
-        return variationId;
-    }
-}
-
-class PercentageRule {
-    @SerializedName(value = "v")
-    private JsonElement value;
-    @SerializedName(value = "p")
-    private double percentage;
-    @SerializedName(value = "i")
-    private String variationId;
-
-    public JsonElement getValue() {
-        return value;
-    }
-
-    public double getPercentage() {
-        return percentage;
-    }
-
-    public String getVariationId() {
-        return variationId;
-    }
-}
